Move vec-cycle test
